mula/requirements-dev.txt
-r requirements.txt
factory_boy==3.2.1
mypy==0.991
black==22.10.0
pylint==2.15.6
click==8.0.4
psutil==5.9.4
